Refactor multi autocomplete ux
