+++ src/components/MultiAutocompleteSelectField/index.ts
@@ -1,2 +1,3 @@
 export { default } from "./MultiAutocompleteSelectField";
 export * from "./MultiAutocompleteSelectField";
+export * from "./MultiAutocompleteSelectFieldContent";
